+++ Runtime/PlatformMovement.cs
@@ -58,8 +58,13 @@ namespace Pixygon.Micro {
                 velocity = new Vector2(velocity.x, _jumpPower);
                 _coyoteTime = 0f;
                 _jumpBuffer = 0f;
-            } else if (started && _coyoteTime > 0f) {
+            } else if (started && !IsGrounded && _coyoteTime > 0f) {
                 Debug.Log("Coyote-time jump!");
+                velocity = new Vector2(velocity.x, _jumpPower);
+                _coyoteTime = 0f;
+                _jumpBuffer = 0f;
+            } else if (started && !IsGrounded && _jumpBuffer > 0f) {
+                Debug.Log("Jump-buffer jump!");
                 velocity = new Vector2(velocity.x, _jumpPower);
                 _coyoteTime = 0f;
                 _jumpBuffer = 0f;
@@ -96,7 +101,6 @@ namespace Pixygon.Micro {
                 _landSfx.Play();
                 _anim.Land();
                 if (_jumpBuffer > 0f) {
-                    Debug.Log("Jump-buffer jump!");
                     Jump(true);
                 }
             }